requirements.txt
streamlit==1.14.1
pandas==1.4.1
numpy==1.23.4
matplotlib==3.5.1
seaborn==0.11.2
tensorflow==2.8.0
keras==2.8.0
glob2==0.7
Pillow==9.0.1
split-folders==0.5.1
pillow
requests